@@ -253,9 +248,8 @@
     "core-js": "^3.21.1",
     "docusaurus-graphql-plugin": "^0.9.0",
     "esbuild": "0.19.5",
-    "eslint": "~8.46.0",
+    "eslint": "8.57.0",
     "eslint-config-prettier": "9.0.0",
-    "eslint-config-react-app": "^7.0.0",
     "eslint-import-resolver-typescript": "^3.5.3",
     "eslint-plugin-import": "^2.27.5",
     "eslint-plugin-jsx-a11y": "6.7.1",
