@@ -1,4 +1,5 @@
-module.exports = {
+/* eslint-disable */
+export default {
   displayName: "back",
   preset: "../jest.preset.js",
   testEnvironment: "node",
